파일 업로드

+++ src/main/java/com/project/groupware/domain/DocumentFileVO.java
@@ -4,14 +4,14 @@ public class DocumentFileVO {
 	private String id;					//고유 아이디
 	private String originalFileName;	//파일 원본명
 	private String systemFileName;		//시스템 파일명
-	private int fileSize;				//파일 크기
+	private long fileSize;				//파일 크기
 	private String documentId;			//문서 아이디
 	
 	public DocumentFileVO() {
 
 	}
 
-	public DocumentFileVO(String id, String originalFileName, String systemFileName, int fileSize, String documentId) {
+	public DocumentFileVO(String id, String originalFileName, String systemFileName, long fileSize, String documentId) {
 		this.id = id;
 		this.originalFileName = originalFileName;
 		this.systemFileName = systemFileName;
@@ -43,11 +43,11 @@ public class DocumentFileVO {
 		this.systemFileName = systemFileName;
 	}
 
-	public int getFileSize() {
+	public long getFileSize() {
 		return fileSize;
 	}
 
-	public void setFileSize(int fileSize) {
+	public void setFileSize(long fileSize) {
 		this.fileSize = fileSize;
 	}
